refactor(portal): extract shared AuthGuard to portal-shared package
