+++ apps/clinician-portal/src/lib/auth-guard.tsx
@@ -1,29 +1,3 @@
 "use client";
 
-import { useEffect } from "react";
-import { useRouter } from "next/navigation";
-import { useAuth } from "./auth";
-
-/**
- * Wraps a page component and redirects to /login if not authenticated.
- */
-export function AuthGuard({ children }: { children: React.ReactNode }) {
-  const { isAuthenticated, hydrated } = useAuth();
-  const router = useRouter();
-
-  useEffect(() => {
-    if (hydrated && !isAuthenticated) {
-      router.replace("/login");
-    }
-  }, [hydrated, isAuthenticated, router]);
-
-  if (!hydrated || !isAuthenticated) {
-    return (
-      <div style={{ padding: 40, color: "var(--text-muted)" }}>
-        {hydrated ? "Redirecting to login..." : "Loading..."}
-      </div>
-    );
-  }
-
-  return <>{children}</>;
-}
+export { AuthGuard } from "@carebridge/portal-shared/auth-guard";
