+++ packages/portal-shared/package.json
@@ -6,7 +6,8 @@
   "exports": {
     "./trpc": "./src/trpc.ts",
     "./auth": "./src/auth.tsx",
-    "./providers": "./src/providers.tsx"
+    "./providers": "./src/providers.tsx",
+    "./auth-guard": "./src/auth-guard.tsx"
   },
   "dependencies": {
     "@trpc/client": "^11.0.0",
@@ -15,9 +16,13 @@
     "@tanstack/react-query": "^5.60.0",
     "react": "^19.0.0"
   },
+  "peerDependencies": {
+    "next": ">=14.0.0"
+  },
   "devDependencies": {
     "@carebridge/api-gateway": "workspace:*",
     "@types/react": "^19.0.0",
+    "next": "^15.1.0",
     "typescript": "^5.7.0"
   }
 }
